+++ src/components/CalendarPage/Calendar.css
@@ -22,9 +22,8 @@
   color: #aaaaaa;
   display: flex;
   justify-content: center;
-  /* margin-top: 1rem; */
 }
-.calender {
+.calendar {
   width: 95%;
   height: 90%;
   display: flex;
@@ -127,7 +126,7 @@
 }
 
 /*날짜(바디부분)*/
-.calenderbody {
+.calendarbody {
   width: 100%;
   height: 89%;
   padding-left: 5px;
